fix: remove shouldLog from public
