+++ packages/log/ts/index.ts
@@ -1,8 +1,7 @@
 export * from './console/index.js'
 export * from './standardLog.js'
 export * from './errors.js'
-export * from './formatter.js'
+export * from './reporter.js'
 export * from './logLevels.js'
 export * from './memory.js'
-export * from './shouldLog.js'
 export * from './types.js'
